Implemented initial methods of the library
This class should be the only public class that let the user access the APIs of the library.
Those 2 static methods are the only way to initial an new UCClient instance, and the "static init()" method utilised lazy initialisation & double-check locking idiom to ensure only one UCClient instance will be initialisated.